Tree-shake manifest validation out of production client bundles
resolveApp re-ran full manifest validation in the browser on every page
load — unknown shell/middleware name checks, loaderCache validation,
SPA+hydration conflict checks, and the did-you-mean error formatting they
reference — even though pracht build already validated the manifest.

Split the pure matching/path/href primitives into route-matching.ts (no
resolveApp or name-suggestions imports) and point the client router,
prefetch listeners, and hooks at it; the router previously reached
resolveApp through matchAppRoute's resolve fallback even though it always
holds an already-resolved app. Guard resolveApp's validation behind
import.meta.env.DEV so production Vite builds constant-fold it away while
dev servers, vitest, and pracht build (Node, where import.meta.env is
undefined) keep full validation.

Measured on examples/basic: framework entry chunks 22.62 → 20.62 kB raw
(8.73 → 7.96 kB gzip); no validation/suggestion strings remain in any
production client chunk. Public API unchanged.
